% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conjoint.R
\name{designConjoint}
\alias{designConjoint}
\title{Create conjoint analysis experimental design.}
\usage{
designConjoint(nLevels = c(4, 3, 2, 1), extraTrials = 3,
  levelNames = LETTERS[1:length(nLevels)], ...)
}
\arguments{
\item{nLevels}{A numeric vector specifying the number of levels for each attribute}

\item{extraTrials}{the number of additional trials to use in experiment}

\item{levelNames}{A character vector with Attribute names}

\item{...}{Other arguments passed to \code{\link[AlgDesign]{optFederov}}}
}
\description{
Create conjoint analysis experimental design.
}
\examples{
designConjoint(c(6, 6, 3, 2))
}
