% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxdiff.R
\name{designMaxdiff}
\alias{designMaxdiff}
\title{Create optimal block design for maxdiff scaling}
\usage{
designMaxdiff(nAttributes, AttrNames = LETTERS(1:nAttributes), choices, sets,
  versions = 1, nrepeats = 100)
}
\arguments{
\item{nAttributes}{The number of attributes to be tested}

\item{AttrNames}{A character vector with names for each attribute.}

\item{choices}{The number of choice tasks in each screen (typically about 4)}

\item{sets}{The desired number of sets (i.e. screens in market research terms)}

\item{versions}{The number of different versions in the design.  Each respondent will only see a single version, but the versions will be randomised across different respondents.}

\item{nrepeats}{The number of time the algorithm will repeat to find an optimal solution}
}
\description{
This is a wrapper around \code{\link[AlgDesign]{optBlock}} to create partially balanced incomplete block designs.
}
