% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalblock.R
\name{orthogonalBlock}
\alias{orthogonalBlock}
\title{Finds an orthogonal design for a given PBIB design.}
\usage{
orthogonalBlock(x, versions = nlevels(factor(x$version)),
  iter = nrow(x)/length(unique(versions)) + 2, restart = TRUE)
}
\arguments{
\item{x}{A partially balanced incomplete block (PBIB) experimental design}

\item{versions}{Number of versions of design}

\item{iter}{Number of iterations}

\item{restart}{TRUE or FALSE}
}
\description{
Finds an orthogonal design for a given PBIB design.
}
\examples{
#Create test data
library(AlgDesign)
set.seed(12345)
choices <- 4
nAttributes <- 7
blocksize <- 7
bsize <- rep(choices, blocksize)
# Find optimal block design
PBIB <- optBlock(~., withinData=factor(1:nAttributes), blocksizes=bsize)
df <- data.frame(t(array(PBIB$rows, dim=c(choices, blocksize))))
colnames(df) <- paste("Item", 1:choices, sep="")
rownames(df) <- paste("Set", 1:nAttributes, sep="")
print(df)
set.seed(12345)
orthogonalBlock(df)
}
\author{
Joris Meys
}
